make puppet-version a layer option
Puppet 3 vs 4 is a massive upgrade. It would be easy to break lots of software
(e.g. Bigtop) if a user inadvertently changed the major puppet version of a
Bigtop charm post deployment. Move the puppet version to a layer option, which
is only mutable at deploy-time.

There is still the 'pin-puppet' config option if users really want to override
our default.